requirements.txt
scipy>=1.3.0
numpy==1.17.3
pandas
sklearn
matplotlib
keras=2.3.1
pyyaml
tqdm
tensorflow==1.13.1
pydot # to plot model using keras
# sudo apt install python-pydot python-pydot-ng graphviz